\name{bdfun0}
\alias{bd0}
\alias{bdfun0}
\title{basic exponential-fecundity gradient function}
\usage{
bdfun0(t, y, parms)

bd0(timeMax = 20, steps = 50, N0 = 1, parms = c(b0 = 1, d0 = 0.5, bDD =
  10))
}
\arguments{
  \item{t}{time}

  \item{y}{state variables}

  \item{parms}{parameter values}

  \item{timeMax}{maximum time}

  \item{steps}{number of time steps}

  \item{N0}{starting value}

  \item{parms}{named parameter vector}
}
\description{
basic exponential-fecundity gradient function

basic exponential-fecundity simulation
}

